refactor: blog layout components
