--- apps/webapp/components/routes/blog/base/index.ts
@@ -1,9 +1,0 @@
-export * from "./button.component"
-export * from "./input.component"
-export * from "./select.component"
-export * from "./alert.component"
-export * from "./checkbox.component"
-export * from "./toggle.component"
-export * from "./range.component"
-export * from "./tag.component"
-export * from "./lazy-image.component"
