--- apps/webapp/components/routes/blog/base/toggle.component.tsx
@@ -1,14 +1,0 @@
-export const Toggle = (props: any) => {
-  const { label } = props
-  return (
-    <label className="relative inline-flex items-center cursor-pointer">
-      <input type="checkbox" className="sr-only peer" {...props} />
-      <div className="transition duration-400 ease-in-out  hover:-translate-y-0.5 hover:scale-105   w-11 h-6 bg-gray-200 peer-focus:outline-none peer-focus:ring-4 peer-focus:ring-primary-300 dark:peer-focus:ring-primary-800 rounded-full peer dark:bg-gray-700 peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-5 after:w-5 after:transition-all dark:border-gray-600 peer-checked:bg-primary-500"></div>
-      {label ? (
-        <span className="ml-3 text-sm font-medium text-gray-900 dark:text-gray-300">
-          {label}
-        </span>
-      ) : null}
-    </label>
-  )
-}
