--- apps/webapp/components/routes/blog/modules/index.tsx
@@ -1,6 +1,0 @@
-export * from "./dropdown-menu.component"
-export * from "./app-bar-dropdown.component"
-export * from "./theme-toggle.component"
-export * from "./main-nav.component"
-export * from "./dialog-component"
-export * from "./header-seo.component"
