--- apps/webapp/components/routes/blog/molecules/index.ts
@@ -1,8 +1,0 @@
-export * from "./hero-sub-card.component"
-export * from "./hero-card.component"
-export * from "./post-card.component"
-export * from "./top-dropdown-component"
-export * from "./mobile-menu-component"
-export * from "./navigator.component"
-export * from "./app-search-bar.component"
-export * from "./locale-modal.component"
