--- apps/webapp/components/routes/blog/views/blogpage/index.ts
@@ -1,2 +1,0 @@
-export * from "./blog.component"
-export * from "./blog.types"
