--- apps/webapp/components/routes/blog/views/category/index.ts
@@ -1,2 +1,0 @@
-export * from "./category.component"
-export * from "./category.types"
