--- apps/webapp/components/routes/blog/views/home/home.types.tsx
@@ -1,23 +1,0 @@
-import { CMSLayoutText, CMSPageSeoText } from "@/services/datocms"
-import {
-  BlogArticleRecord,
-  getBlogCategoriesTypes,
-} from "@/services/datocms/datacms-blog-category.service"
-import { LayoutRecord } from "@/services/datocms/graphql/generated/cms"
-
-export interface HomeProps {
-  children?: React.ReactNode
-  i18n: CMSLayoutText
-  sectionArticles: {
-    name: string
-    slug: string
-    articles: BlogArticleRecord[]
-  }[]
-
-  recentsArticles: {
-    name: string
-    slug: string
-    articles: BlogArticleRecord[]
-  }[]
-  pageSeo: CMSPageSeoText
-}
