--- apps/webapp/components/routes/blog/views/home/index.ts
@@ -1,2 +1,0 @@
-export * from "./home.component"
-export * from "./home.types"
