--- apps/webapp/components/routes/blog/views/page/index.ts
@@ -1,2 +1,0 @@
-export * from "./page.component"
-export * from "./page.types"
