--- apps/webapp/components/routes/blog/views/page/page.types.tsx
@@ -1,11 +1,0 @@
-import {
-  CMSLayoutText,
-  CMSPageSeoText,
-  MainContentBlock,
-} from "@/services/datocms"
-
-export interface TermsPageProps {
-  i18n: CMSLayoutText
-  pageSeo: CMSPageSeoText
-  mainContent: MainContentBlock | null
-}
